Refactor device.py into a few modules
