+++ tests/test_device_sync.py
@@ -67,41 +67,30 @@ def _patch_micropython_code(snippet):
 
 @pytest.fixture
 def sync_begin():
-    snippet = belay.device._read_snippet("sync_begin")
+    snippet = belay.device.read_snippet("sync_begin")
     snippet = _patch_micropython_code(snippet)
     exec(snippet, globals())
 
 
 @pytest.fixture
 def hf():
-    snippet = belay.device._read_snippet("hf")
+    snippet = belay.device.read_snippet("hf")
     snippet = _patch_micropython_code(snippet)
     exec(snippet, globals())
 
 
 @pytest.fixture
 def hf_native():
-    snippet = belay.device._read_snippet("hf_native")
+    snippet = belay.device.read_snippet("hf_native")
     snippet = _patch_micropython_code(snippet)
     exec(snippet, globals())
 
 
 @pytest.fixture
 def hf_viper():
-    snippet = belay.device._read_snippet("hf_viper")
+    snippet = belay.device.read_snippet("hf_viper")
     snippet = _patch_micropython_code(snippet)
     exec(snippet, globals())
-
-
-def test_sync_local_belay_hf(tmp_path):
-    """Test local FNV-1a hash implementation.
-
-    Test vector from: http://www.isthe.com/chongo/src/fnv/test_fnv.c
-    """
-    f = tmp_path / "test_file"
-    f.write_text("foobar")
-    actual = belay.device._local_hash_file(f)
-    assert actual == 0xBF9CF968
 
 
 def test_sync_device_belay_hf(hf, tmp_path):
@@ -209,7 +198,7 @@ def test_device_sync_partial_remote(mocker, mock_device, sync_path):
             if local_fn.stem.endswith("1"):
                 out.append(0)
             else:
-                out.append(belay.device._local_hash_file(local_fn))
+                out.append(belay.device.fnv1a(local_fn))
         return out
 
     def mock_exec(cmd, data_consumer=None):
